+++ lib/updater_controller.dart
@@ -2,14 +2,22 @@ import "package:desktop_updater/desktop_updater.dart";
 import "package:flutter/material.dart";
 
 class DesktopUpdaterController extends ChangeNotifier {
-  DesktopUpdaterController({required Uri? appArchiveUrl, this.localization, this.skipCheckVersion = false}) {
+  DesktopUpdaterController({
+    required Uri? appArchiveUrl,
+    this.localization,
+    bool skipInitialVersionCheck = false,
+    @Deprecated("Use skipInitialVersionCheck instead.") bool? skipCheckVersion,
+  }) : _skipInitialVersionCheck = skipCheckVersion ?? skipInitialVersionCheck {
     if (appArchiveUrl != null) {
       init(appArchiveUrl);
     }
   }
 
-  bool skipCheckVersion;
-  bool get getSkipCheckVersion => skipCheckVersion;
+  final bool _skipInitialVersionCheck;
+  bool get skipInitialVersionCheck => _skipInitialVersionCheck;
+
+  @Deprecated("Use skipInitialVersionCheck instead.")
+  bool get getSkipCheckVersion => _skipInitialVersionCheck;
 
   DesktopUpdateLocalization? localization;
   DesktopUpdateLocalization? get getLocalization => localization;
@@ -64,10 +72,13 @@ class DesktopUpdaterController extends ChangeNotifier {
 
   void init(Uri url) {
     _appArchiveUrl = url;
-    if (!skipCheckVersion) {
-      checkVersion();
+    if (_skipInitialVersionCheck) {
       notifyListeners();
+      return;
     }
+
+    checkVersion();
+    notifyListeners();
   }
 
   void makeSkipUpdate() {